+++ js/util.js
@@ -1,0 +1,68 @@
+// Функция тосавания массива при помощи алгоритма Фишера-Йетса. Взята из интернета и доработана
+// Источник - https://sebhastian.com/fisher-yates-shuffle-javascript/
+function fyShuffle(arr) {
+  let i = arr.length;
+  while (--i > 0) {
+    const randIndex = getRandomPositiveInteger(0, i);
+    [arr[randIndex], arr[i]] = [arr[i], arr[randIndex]];
+  }
+  return arr;
+}
+
+// Функция взята из интернета и доработана
+// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+export function getRandomPositiveInteger (a, b) {
+  // Если переданы отрицительные числа, возвращаем NaN
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+
+  // Чтобы не заставлять пользователя нашей функции помнить порядок аргументов,
+  // реализуем поддержку передачи минимального и максимального значения в любом порядке,
+  // а какое из них большее и меньшее вычислим с помощью Math.min и Math.max.
+
+  // После нам нужно убедиться, что пользователь не передал дробные значения,
+  // для этого на всякий пожарный случай нижнюю границу диапазона
+  // мы округляем к ближайшему большему целому с помощью Math.ceil,
+  // а верхнюю границу - к ближайшему меньшему целому с помощью Math.floor
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  // Дальше используем Math.random() для получения случайного дробного числа в диапазоне [0, 1),
+  // которое домножаем на разницу между переданными числами плюс единица - это будет наша случайная дельта.
+  // После нужно сложить дельту с минимальным значением, чтобы получить итоговое случайное число.
+  const result = Math.random() * (upper - lower + 1) + lower;
+  // "Плюс единица", чтобы включить верхнюю границу диапазона в случайные числа
+
+  // И в конце с помощью метода Math.floor мы округляем полученный результат,
+  // потому что Math.random() генерирует только дробные числа и ноль.
+  return Math.floor(result);
+}
+
+// Функция взята из интернета и доработана
+// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+export function getRandomPositiveFloat (a, b, digits = 1) {
+  // Если переданы отрицительные числа, возвращаем NaN
+  if (a < 0 || b < 0 || digits < 0) {
+    return NaN;
+  }
+
+  // Чтобы не заставлять пользователя нашей функции помнить порядок аргументов,
+  // реализуем поддержку передачи минимального и максимального значения в любом порядке,
+  // а какое из них большее и меньшее вычислим с помощью Math.min и Math.max
+  const lower = Math.min(a, b);
+  const upper = Math.max(a, b);
+
+  // Дальше используем Math.random() для получения случайного дробного числа в диапазоне [0, 1),
+  // которое домножаем на разницу между переданными числами - это будет наша случайная дельта.
+  // После нужно сложить дельту с минимальным значением, чтобы получить итоговое случайное число.
+  const result = Math.random() * (upper - lower) + lower;
+
+  // И в конце с помощью метода toFixed любого числа в JavaScript
+  // указать требуемое количество знаков после точки.
+  // Метод возвращает строку, поэтому с помощью унарного плюса превращаем её в число
+  return +result.toFixed(digits);
+}
+
+export const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+export const getShuffletArrayWithRandomLength = (arr) => fyShuffle([...arr]).slice(getRandomPositiveInteger(0, arr.length));
